Fix a bug for SVGTransformList's replace function (#3613)
In SVGTransformList, when an item is removed or replaced, the item first
be erased. The erase functions support only two types argument
:size_t and template<>. But, in 32 bit machine, unsigned long type
variable is recognized to template<>, so the different results happen.
Therefore, this patch fixes the bug and adds toString for SVGTransform.
In addition, this adds a testcase for TV target.

+++ src/core/dom/svg/SVGTransformList.h
@@ -81,7 +81,7 @@ protected:
     bool removeItemWithoutUpdateAttribute(unsigned long index)
     {
         if (index < length()) {
-            m_v.erase(index);
+            m_v.erase((size_t)index);
             return true;
         } else {
             return false;
